requirements.txt
Flask~=3.0.0
flask_restful
python-dotenv
mysql-connector-python
flask-cors